Uncap amqp 2.0+ in global requirements
1.x is not compatible with a newer kombu 4.0.0 released today. It's
either we cap kombu to <4.0.0, or uncap amqp. The former is proposed as:
I7b8d467ef612b1023fb2b680dd222ad75f494525 This patch is an alternative
solution that should hopefully allow us to move forward with upstream
amqp/kombu releases.

This patch also bumps versions of the libraries in upper-constraints.txt
to the latest to see if they fail in cross-check gate jobs for any
covered projects.

Change-Id: I70c2d04561abf6e66b6e1ec27df33b289c9c7698
Closes-Bug: #1638263