Update patch set 2
Patch Set 2:

> is 1.15 the one that unifies the py2 and py3 versions?

Yup.  As of 1.14.0 dnspython has a single code base for both py2 and py3 so setting the minimum and switching projects (swift and designate*) to use dnspython rather than a mix is the end game here.

the converged tree also adds new features so we want to tread a little carefully but that's the plan

Patch-set: 2
Label: Verified=0