Update patch set 2
Patch Set 2:

> @Dirk @Tony it's really not cool sneak-merging https://review.openstack.org/381780...
 > Now we're bound to older version of docker-py than necessary. And
 > tripleo-common will never get their requirements updated because
 > they are updated from this repo.

Sorry you feel that way.  The core team evaluated the options and made a decision that best reflects the mission of the requirements team.

Yes it's a step back, yes that frustrating.  However as stated when we reverted the docker-py change we *can* remove the 1.8.0 cap once the pull request is merged and released.

Once that happens we can and will work with the tripleo team to update and release an updated version of tripleo-common.

Patch-set: 2