Update patch set 3
Patch Set 3:

recheck

This one worries me a little.  I can't explain how this constraints only change can affect keyston (which doesn't use the package in question) and yet this is the only review in the last 7 days to hit this exception so the tests doesn't seem to be at fault.

Patch-set: 3