Bump os-client-config to 1.22.0
A number of releases toward the end of Newton caused trouble
for the consuming projects.  The beginning of Ocata is a good time
to bump everyone up to the current happy-fun-goodtime release.

This will also allow OSC and osc-lib to get rid of a bunch of
compatibility-workaround code without harming shade or other
consumers of o-c-c.

Change-Id: I296b6eee97affd6bbacf48e91a5db1b01ddebb4b
Depends-On: I79e2416d452a937b956f3a5680c25c968f6b6aaf