Update patch set 1
Patch Set 1: Code-Review+1

Traditionally we've recommended the use of mod_wsgi, but Octavia needs this to run on their service-vms which are minimal and definitely shouldn't be running something as heavy as apache2...
gunicorn provides an api to run a simple lightweight wsgi listener to be used internally by the Octavia service. Currently the amphora uses flask's werkzeug to serve, but it has a number of bugs that cause instability in practical usage.

Patch-set: 1
Label: Code-Review=+1