Update patch set 1
Patch Set 1: Code-Review-1

I'd rather not set this precedent.  I think a keeping octavia's requirements current while allowing for gunicorn is the right way forward.

Of course I may regret that decision in the future but a good aspect of that regret is it's easy to reverse.

Patch-set: 1
Label: Code-Review=-1