Update patch set 1
Patch Set 1:

> Tony, your comment makes it sound like you are regretfully +1, but
 > you did -1? Maybe I don't quite understand what you said.

Sorry my comments are confusing.  I meant  the -1.  my double talk about regrets was more ... We do a bunch of work to make this function[1] and then realise we overlooked something.

At that point we we can reverse the decision to exclude it.  BUT if we add it now it's harder to exclude it later.

Of course since leaving that review the mailing list discussion has moved on and I now need to re-evaluate

[1] Modify tools, files and jobs as outlined in Options 2 or 3 in the mailing list thread

Patch-set: 1