Remove upper limit for pylxd to 2.1.1
pylxd 2.1.2 removes a backwards-compatible reference
to deprecated code. Upstream thought it was a good idea,
until the tests for stable/newton of nova-lxd began breaking.

Unfortunately, in order to fix the incompatibility, we had to pin
2.1.1 in stable/newton. When looking at stable/newton of requirements,
however, it's WAY out of date for what master had when newton was released.
This patch looks like we're adding versions to upper constraints, when
we're really updating requirements to reflect what nova-lxd already
supports.

Fixes bug 1638037

Change-Id: I1ddb752d4045d040f597683696013e8a4eadf741