Update patch set 2
Patch Set 2: Code-Review-1

I'm still confused.

Is the amphora-agent a "core" part of octavia, i.e something that every deployer will need to use?  Or is it a tool used by the octavia project for verifying correct operation in the gate?

My impression from the mailing list discussion was the later (used only in the gate) but looking at octiavia's setup.py it seems like everyone gets it.

If amphora-agent is only used in the service VM and then only in the gate then I feel like adding this to global-requirements is the wrong answer, as it doesn't need to be co-installable.

I know it's bad form to -1 with a question but it's a pretty fundamental question.

Patch-set: 2
Label: Code-Review=-1