Update patch set 2
Patch Set 2:

I don't blame you for being confused -- it's a bit confusing.
So, the amphora-agent is something everyone needs to use, BUT it is just a WSGI app that can be deployed with any WSGI runner (gunicorn, uwsgi, werkzeug, etc) in the amphora image the deployer builds.
In our gate job, we build our test amphora image using gunicorn.
It is "hardcoded" in a sense, in agent.py -- but while we use agent.py as our binary, a deployer could deploy without using that, instead by setting up any runner to load our WSGI app directly. Or, a deployer could use our default gunicorn setup, but either way it has no co-installability requirement as in any case it runs inside a service VM.

I'd say we're opinionated but WSGI-runner agnostic at the same time.
Adding gunicorn to requirements makes it significantly easier to track our dependency on it, and feels like "the right thing to do", though it is not strictly necessary and we could just include it directly in our DIB Element for installing the agent. Internally we had pushback against that approach, but if you guys think we should go that route (which sounds like it has been suggested), we can revisit it.

One last note: I keep hearing things about "we don't want to use pecan" because it is heavweight or something (possibly not here, but on the ML and the other CR), but pecan is a WSGI framework not a runner -- we could swap out flask for pecan if people really wanted us to, but we'd still need gunicorn to RUN the pecan app, so pecan is orthogonal to this discussion.

Patch-set: 2