Update patch set 2
Patch Set 2: Code-Review+1

About the coinstallability piece, if gunicorn is supported as one of wsgi runners that openstack supports I think it should be in requirements.  The reason I think so is because unlike ruby, we do need to maintain co-installability of python libraries.  Since gunicorn is a python library it should be tracked.  The other impls of wsgi runners don't need it I think, uwsgi and apache-wsgi don't have python libraries.

Patch-set: 2