Update patch set 6
Patch Set 6:

So, I'm happy to change g-r back, but I'd like to register my concern that the released g-r is absolutely incorrect. There is not a released version of LXD that works with a version of pylxd that old. Normally, when there is a bug in the released version of something, we go back and fix the bug, not say "well, we can't do fix this bug because policy/review/whatever". There is a bug in g-r, and I tried to fix it.

Patch-set: 6
Reviewer: Gerrit User 7658 <7658@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0