Update patch set 2
Patch Set 2:

> seems that jenkins didn't apply the Depends-On commit beforehand.

It can't *apply* a depends-on the way you'd expect as you're depending on a gate config change which can't be done on the fly per job.  The Depends-On still has value as it behaves as you'd expect WRT to blocking the merge.  Also it provides clear documentation.

Patch-set: 2