Add sphinxcontrib-fulltoc
This library will be used for docs building by git-upstream.

The library is a relatively simple extension for the Sphinx
documentation system already in use for OpenStack documentation
generation.

It enables a combination of the global table-of-contents and local
table-of-contents to appear in the sidebar of the generated
documentation making it easier to navigate docs by providing access to
all other top level topics in the sidebar when on any page, rather than
forcing the user to go back to the top level index each time.

While it's not actively changing, there also doesn't appear to be much
need for it to be modified as it currently works for the common use
cases expected.

The license (Apache 2.0) is compatible.

Change-Id: I8dced20b036c267799b609a6fa719f6afde89a10