Update patch set 3
Patch Set 3: Code-Review-1

So we just have a tripleo-common release that is queued for uc, version 5.3.0. we could probably get a 5.3.1 with the relaxed g-r.

I suggest to solve this cleanly and first only merge the g-r change. Hongbin, if you agree, please upload a new revision of this change only touching global-requires. The uc change will come automatically once we got the tripleo-common change in



I

Patch-set: 3
Label: Code-Review=-1