Update patch set 2
Patch Set 2:

We now have a fix to make it backward compatible[1].

The change was to fix a bug, where sahara[2], which was using endpoint_override(which is the correct way to override an endpoint with sessionclient).

May be we can have a release quickly to avoid the blacklist?

[1] https://review.openstack.org/#/c/400195/
[1] https://bugs.launchpad.net/python-heatclient/+bug/1621505/comments/2

Patch-set: 2