Update patch set 2
Patch Set 2: Workflow+1

I think I understand this now and whiles I'm still on the side of this shouldn't be in global-requirements my pragmatic side just says it's easiest for octavia and it isn't horrendous so ....

My -1 is to capture that I disagree with this but I'll approve it as it unblocks octavia.

I'd like to find a way to signal that this isn't really for general consumption and the only way I can think to do that is by adding a block comment in global-requirements.txt but that could easily be missed by consumers.

Patch-set: 2
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=+1