Update patch set 1
Abandoned

Actually - this might not be required yet. I misunderstood the api docs for passlib and i was trying to use the `using()` method which ins't included in passlib 1.6.5. That should be fine because the .hash() and .encrypt() methods both are, so we should be able to get around the deprecation warning the bug without bumping the version.

Patch-set: 1
Status: abandoned