requires python-kafka>=1.3.1
oslo.messaging kafka driver development is blocked since 6 months
due to the kafka cap.

The Monasca performance issue when the deprecated API is used have been
fixed in 1.3.1.

This change also removes the monasca libs since they are not compatible
with the new kafka.

So this change remove the cap and raise the minimal requirement.

http://lists.openstack.org/pipermail/openstack-dev/2016-December/108350.html

Change-Id: I5aa71d1bd6f741f533e20e4ccf5f75a7b7ef3fac