Add neobunch
Is the library actively maintained?
Yes. Last update was on November 2015.
https://pypi.python.org/pypi/neobunch

Is the library python 3 compatible?
Yes. The code is python 3 compatible.

Is the library license compatible?
Yes. The license is MIT.

Is the library already packaged in the distros we target?
The older bunch package is available in Ubuntu and Fedora.
Available for ubuntu at:
http://packages.ubuntu.com/search?keywords=python-bunch
Available for Fedore at:
https://apps.fedoraproject.org/packages/python-bunch
The neobunch is not available for these distros.

Is the function of this library already covered by other libraries
in global-requirements.txt?
Not that I could see.

Is the library required for OpenStack project or related dev or
infrastructure setup?
Yes. Required for Cinder and specifically for the following
review: https://review.openstack.org/#/c/404958/

Change-Id: I3b890f7bf2187f614298fdcd0b14b03be3ee1737