Add bunch
Is the library actively maintained?
The recent updates were quite some time back, but the code is very stable and
there was no need to update it. https://pypi.python.org/pypi/bunch/1.0.1

Is the library python 3 compatible?
Yes. The code is python 3 compatible.

Is the library license compatible?
Yes. The license is MIT.

Is the library already packaged in the distros we target?
The library is available for ubuntu at:
http://packages.ubuntu.com/search?keywords=python-bunch

Is the function of this library already covered by other libraries in
global-requirements.txt?
Not that I could see.

Is the library required for OpenStack project or related dev or infrastructure
setup?
Yes. Required for Cinder.

Change-Id: I3b890f7bf2187f614298fdcd0b14b03be3ee1737