Update patch set 1
Patch Set 1: Code-Review-1

I don;t think this is a good idea.  openstack-requirements shoudl not be used in any code we "ship".  Is the and example of a consumer that *needs* this in g-r and u-c?

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1