Update patch set 5
Patch Set 5: Code-Review-1

(1 comment)

Please show is the POC patch.

I'm on the fence here, this seems like a terrible idea but from a requirements POV it's fine.  So I'll err on the side of "Looks good to me" once there is a POC patch against cinder.

Patch-set: 5
Label: Code-Review=-1