Update patch set 6
Patch Set 6: Code-Review-1

It is not tagged as Python 3.x compatible on pypi, although that shouldn't be a deal breaker if it indeed supports python 3.x the last release is from 2011 which doesn't give a good feeling in that regard though. there is also trivial stuff open for months: https://github.com/dsc/bunch/pull/37

given the ultra minimal additional value of this library, would it be tough to rework that code that uses Bunch? its "just" for json or yaml serialisation in this case? didn't read the review in detail yet.

Patch-set: 6