Update patch set 3
Patch Set 3: Code-Review+1

i don't really agree with this. oslo.messaging is not used by monasca but it's used by basically every other project in OpenStack. it seems like we're blocking all of openstack just to satisfy a single project which doesn't actually use it? wouldn't this be better to just allow monasca to set it's own requirements and allow the rest of openstack to follow these?

Patch-set: 3
Label: Code-Review=+1