Update patch set 3
Patch Set 3:

> i don't really agree with this. oslo.messaging is not used by
 > monasca but it's used by basically every other project in
 > OpenStack. it seems like we're blocking all of openstack just to
 > satisfy a single project which doesn't actually use it? wouldn't
 > this be better to just allow monasca to set it's own requirements
 > and allow the rest of openstack to follow these?

Gord, Monasca does use python-kafka and have reasonable technical issues with the latest library.  Of course the Monasca team can elect to opt-out of requirements management the same way the Telemtry have but right now that's not a route they've chosen to take.

It is reasonable to block another teams development for 8+months?  That's an entirely different question.

As stated my -2 is because I *know* this change will break consumers It's no different to a nova core blocking a feature in libvirt that will break all the Xen comsumers.

Patch-set: 3