Update patch set 3
Patch Set 3:

> It is reasonable to block another teams development for 8+months? 
 > That's an entirely different question.
 > 
 > As stated my -2 is because I *know* this change will break
 > consumers It's no different to a nova core blocking a feature in
 > libvirt that will break all the Xen comsumers.

that's fair. the 8+month block is what i was concerned about.

i understand that we treat requirements in a way to say that all services can run together in an all-in-one but maybe there's a way to transition away from this? conflicts i imagine can be solved via containers, multiple nodes, etc... don't have a solution but i don't know if we've ever thought of moving away from the all-in-one concept or is it because we're so reliant on devstack?

Patch-set: 3