Update patch set 2
Patch Set 2:

Well, other projects typically should use the upper-constraints file, but those that do will not have the problem, since we don't use the broken version in there.

Those that do not would need to explicitly list their dep on docutils (which is imho wrong) and then wait for the proposal bot update. By that time this regression is maybe already fixed by upstream.

Anyway, we still need the change here anyway, so that's good to merge imho.

Patch-set: 2