remove upper cap of kombu/amqp
kumbu/amqp upper caps have been set because their API have changed in
4.0.0/2.0.0

We are working on oslo.messaging side to use the new API [1]. But we
can't test it in gate due to the current version in global requirements
and upper-constraint.

Unfortunately this change can't be merged, because I haven't updated the
upper-constraint. But also if I update upper-constraint, the
upper-constraint will be incompatible with the last released
oslo.messaging.

How should we handle this requirement update ? Do I miss somehting ? Should we remove
upper-constraint from lib ? Something else ?

[1] I73958c8057353a2eefe1baaa7a41148193d507f7

Change-Id: I9f980b51901ac31599b9651633956ad2eea6a1ac