Update patch set 1
Patch Set 1:

That's so many steps to change 3 lines...

This means we have to merge an untested change :(

For this one we are lucky supporting both looks doable. But that will not be always the case.

I'm not a huge fan of local testing, specially for oslo.messaging, we have so many dsvm jobs that ensures we didn't break others. 

Anyway, if that the solution, I will try.

I have another question, if I just remove the upper-cap how to ensure the "
Updated from generate-constraints" bot will not update the upper-constraints ?

Patch-set: 1
Label: Verified=0