Update patch set 2
Patch Set 2: Code-Review-1

As described in [1]  we need the answer to a few questions to help evaluating the addition of this library to the global-requirements.  Please copy the questions ans answers into your commit message:

[Q] Is the library actively maintained?
[A] 

[Q] Is the library good code?
[A] 

[Q] Is the library python 3 compatible?
[A] 

[Q] Is the library license compatible?
[A] 

[Q] Is the library already packaged in the distros we target
    (Ubuntu latest / Fedora latest)?
[A] 

[Q] Is the function of this library already covered by other
    libraries in global-requirements.txt?
[A] 

[Q] Is the library required for OpenStack project or related dev
    or infrastructure setup? (Answer to this should be Yes,
    of course) Which?
[A] 

It's especially helpful if you can include a link to a change in gerrit that needs this library.

[1] http://git.openstack.org/cgit/openstack/requirements/tree/README.rst#n265

Patch-set: 2