Bump SQLAlchemy UC to 1.1.4
When installing devstack and error sometimes occurs running
dbsync on the cinder dtabase:

ERROR oslo_db.sqlalchemy.exc_filters ProgrammingError: (1064, u"You have
an error in your SQL syntax; check the manual that corresponds to your
MySQL server version for the right syntax to use near '(6)' at line 1")

This seems to happen when running SQLAlchemy 1.0.16 but does not occur
with 1.1.4

This change moves the upper-constraints.txt value for SQLAlchemy
to 1.1.4 to remedy thie issue.

Note: There may be an issue with the bot, since the SQLAlchemy page
lists both 1.0.16 and 1.1.4 in its index and the bot chooses the 1.0.x
version.

Change-Id: I48c70d2cec8113f03e4180ed587ad8ee1623061c