Limit Kombu to <4
oslo.messaging is not compatible with Kombu 4 and superior. There's no
point in excluding each micro version of 4.0, which breaks e.g.
Ceilometer gate for several days each time Kombu decide to do a bug fix
release.

oslo.messaging needs to release with a correct indication of what it
supports, and it cannot do it currently because of the content of this
repository.

Change-Id: Ieec53c3ac1d3d9ae69f58f378381d3d94a8ce3ac