Update patch set 1
Patch Set 1: Code-Review+1

This patch is correct.

It's not because oslo.messaging is going to support Kombu 4 *someday* that it should be released now as being marked as supporting it – which is the case since it just blacklisted a few random releases.

So it'd be good to have releases of oslo.messaging which has correct dependency wrt Kombu for once, so we don't have to wait several days and re-release oslo.messaging each time Kombo release a micro version.

Patch-set: 1
Label: Code-Review=+1