Update patch set 1
Patch Set 1:

I agree that we should lock to kombu < 4 on the stable/ branches, but for master I prefer to use constraints to limit upper versions rather than version caps in g-r, because those are harder to get rid of. in this case though there are only few projects depending on kombu directly, so whatever we decide it should be manageable.

Patch-set: 1