Update patch set 1
Patch Set 1: -Code-Review

Sure.

You'll need to get to tweak the constraints generator to pass the each requirement through packaging.canonical_name().

It shoudl be reasonably safe to do this in u-c.txt but not g-r.txt  Also if there is a single place in the suse toolchain/process that could parse names though the function above then that might be slightly easier

Patch-set: 1