Update patch set 1
Patch Set 1:

This is a fair point, but this is more about correcting the issue now while we've got operators in the field who are FUBAR'ed.  

The long term solution I would say would be to ultimately get to these versions (btw Kombu 4.X installs AMQP 2.X which is something we need to go after as well).

Patch-set: 1