Update patch set 1
Patch Set 1:

When Kombu 4 came out and it broke things we (the requirements team) took the approach that blocking individual releases was better than capping as capping tends to last much longer then is desirable and can result in situations like we;re seeing with psutil and python-kafka.

However we didn't (and couldn't reasonably) predict the cadence of kombu 4.X releases.

There are a number of projects NOT using constraints and as such they're only getting half of the benefit of the work the requirements and infra teams put into requirements and constraints management.  We're attempting to address that with https://review.openstack.org/#/q/is:open+topic:feature/add-constraints-support but to be frank the progress hasn't been great.

The main consumer of kombu is oslo.messaging and they're working hard on getting kombu4 supported[1] so I'm in favor of taking a patch like this with an appropriate commit message and (possibly) backporting that to stable.  Then once oslo.messaging is ready they can remove the cap on master

[1] https://review.openstack.org/409294

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=0