Update patch set 1
Patch Set 1:

Hi Tony,

One thing to keep in mind is that all of that work is around "gating", but not about consuming things. They are people using telemetry and oslo projects that are just not about running tests.

Currently a user installing oslo.messaging via pip will grab kombu 4.1.x or 4.0.y that is not blacklisted and it will fail for no good reason. That's why this patch is correct.

The approach taken do not absolutely avoid problem like the one with psutil and kafka; until oslo.messaging supports Kombu 4, patch will kept coming to blacklist them…

So not setting the proper requirement entry is just making life a pain to any user or developer using oslo.messaging.

Patch-set: 1