Update patch set 1
Patch Set 1: Code-Review-1

This looks ok to me but looking at your example implementation it seems to be like you're only using the parts of jsonmodels that can be done with oslo.versionedobjects (ovo).

Befper we add a new library to our ecosystem can you help me understand the ways that ovo is *not* appropriate?  I get that ovo isn't "json backed" but what I don't understand is do you need json backed or is that just a side-effect of the library choice

Patch-set: 1