Update patch set 1
Patch Set 1: Code-Review-1

is there a bugreport about this? updating webob is not a trivial task, there are a lot of failures. I wonder if it would be possible to solve the issue in a way that doesn't require us to bump WebOb immediately.

Patch-set: 1
Reviewer: Gerrit User 6593 <6593@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1