Update patch set 1
Patch Set 1:

Hi Dirk,

Yeah, based on the tests results I can see it's a problem to update it. I might need some help with ideas here because I think we have a chicken-and-egg problem.

The main issue I think is because keystonemiddleware doesn't respect the upper-constraint.txt file [0] [1], so they already test against 1.7.0 and fixed the "charset" problem there. The problem is that the fix broke it for older versions, which is what we use in Ironic.

[0] I opened a bug about it https://bugs.launchpad.net/keystonemiddleware/+bug/1653669

[1] http://logs.openstack.org/02/415502/1/check/gate-keystonemiddleware-python35/b3ca351/console.html.gz#_2016-12-28_15_55_25_458992

Patch-set: 1
Reviewer: Gerrit User 6773 <6773@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0