Add construct lib for data parsing
The python construct lib can be used to facilitate parsing of binary data,
for example network protocols.  It is documented here:
http://construct.readthedocs.io/en/latest/

An example of this library in openstack is in parsing network
switch LLDP data as implemented here: https://review.openstack.org/#/c/406496/.
It simplifies the description of the LLDP structures and makes parsing of bit
and byte fields relatively straightforward.

The construct lib is licensed under the MIT license.  It supports Python 2.7
and 3.5.

This lib appears to be actively maintained, the latest release was on
12/31/16 in the git repository
https://github.com/construct/construct/releases.

Change-Id: I5debba09e832a541b6d139856bb601f31da4a7ba