Update patch set 5
Patch Set 5:

Steve, that redistributor has sent quite a few of that nearly same bug report to other projects (including Hacking).

The problem is that Hacking and newer versions of Flake8 (and pycodstyle) are wildly incompatible due to how Hacking (against my advice) implemented plugins both in itself and for local factories in OpenStack.

I have an idea how to fix this, but haven't had the time to test or implement it and don't want to send others down a rabbit hole.

Patch-set: 5