Update patch set 1
Patch Set 1:

@Tony

We've discussed using OVO as the library for our model back-end but I don't think it fits too well for now.
Our architecture is rather simple, we have distributed key/value database for persistent store, and a pub-sub system for event notification. 
Neutron server part write to DB and publishes an event to relevant topics and the agents subscribe to them and get notified or read other relevant data from the database.

Key problem we want to solve right now is easier model class definition (a-la SqlAlchemy ORM), and uniform treatment of our models in many of the facilities we have in our code. As long as we're accessing the database from our agents we can't really version the objects and we'll have to tread carefully to not break older agents during an upgrade, we're aware of that issue, and will solve it eventually (and then we might want to version our objects).

We can define our models with OVO classes but the output of obj_to_primitive is not exactly the thing we want, (we store objects similar to just versioned_object.data parts of the output) and using OVO with the code that converts to and from this format makes me wonder we're using the right tool for our problem.

[1] https://github.com/openstack/dragonflow/blob/master/dragonflow/db/models.py#L99

Patch-set: 1