Update patch set 2
Patch Set 2: Code-Review-1

I know we talked in person, but here's my response.                             
                                                                                
If jinja2 made a breaking change that won't be fixed in a minor release AND     
ansible can't do a patch/minor release that fixes the issue then the cap is     
needed.  That said, I'm not sure jinja upstream has been engaged enough in this 
to know if they can fix this change in behavior.

Patch-set: 2
Label: Code-Review=-1