bump minimum version of openstackSDK to 0.9.12 (again)
The OSC team addressed the issues that were discovered in the latest
openstackSDK releases (0.9.11 and 0.9.12) with their latest OSC version,
3.7.0.

We are opting to unblock these SDK versions because we do not know when
the SDK team will release 0.9.13.

This reverts commit 6da1fafee9ec30d3d9653c64cd27f61bbfb21862, which
in turn reverted 85277c19c623e42f3c016cec099925d1be713287.

Change-Id: Ib2bb2368459d6f940be59f49fb76bf147c445610