Update patch set 1
Patch Set 1: Code-Review-1

This is probably okay but the fact it's only on goole code worries me a little.

Can someone with more of an interest in this contact the author and see if they'd be interested in reviving the project and/or allowing OpenSTack (monasca) to adopt the library?

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1