Update sphinx requirement
In change I38580a5f765c7bb586dcf2786e91be15a8eba235 we skipped one of the
docutils versions due to a breaking release that was not supported by
sphinx. This has now been fixed in the latter[0].

With Sphinx 1.5.1, the blocking of docutils 0.13.1 is not needed
anymore, so remove the condition. This line could be removed completely,
but for that it first needs to be removed from all projects.

also cherry-pick Correct latex_document settings for Sphinx 1.6.3[1]

[0] https://github.com/sphinx-doc/sphinx/issues/3212
[1] I1fb81ea1d7de561185a6ae31dd4bc73e1ac8f9ae

Change-Id: I01b7b6f3fac75e284debe211129b7b1e87c756d9
(cherry picked from commit 4d8824708b560bf97a63e59786496cc720a95fed)