Update patch set 3
Patch Set 3: Code-Review-1

> I don't konw why the change id links to the commit in master branch
 > and can not find the review of this commit in newton(but it can be
 > found in git log).

Oh don't worry to much about that.  Can you please focus on describing the problem so we can find an acceptable solution.

Raising the minimums on stable branches wont work.  We need to understand which versions are acceptable and we'll need to consider how to make the stable branch of Karbor compatible with 0.0.9

Patch-set: 3
Label: Code-Review=-1