Update patch set 4
Patch Set 4:

Did the update.
In addition i bumped the version to the latest 0.9.0 (which got released this week). The dependencies didn't change.

@Dirk: Regarding: "as it is a optional dep for non-x86"
yes, it is only required if a dpm compute node is present. But the dpm compute node itself can also be x86 system in theory. It can be any system that has access to the ReST API of the system.

I also added more details to the corresponding section in the commit message.

Patch-set: 4