Update patch set 4
Patch Set 4:

matt we change the os-vif vif field to 1.1 to reflact the new field and added the downgrade to 1.0. if the unit test are breaking they were makeing an assertion that they should not.

that said lets try and fix them but they should not break with the introduction of the new field. were they explcitily creating a v 1.0 object?

Patch-set: 4