Update patch set 1
Patch Set 1: Code-Review-1

Current plan is that projects limit tbe max version by using a copy of the upper-constraints.txt file. This works for Jenkins and for those projects that subscribe to it, but it doesn't help for those that just install from pip directly.

More often than not newer versions are not causing a problem, so capping is often unnecessary and also causes great upgrade pain.

Happy to hear a obout ideas on how to make this better but it's not a trivial problem.

Patch-set: 1
Reviewer: Gerrit User 6593 <6593@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1