Update patch set 1
Patch Set 1: Code-Review-1

It sounds like there is a conflict between the oslo.messaging project and the monasca project as to which version of kafka is supported.  Doing (what to me looks like) a run around and removing monasca doesn't solve the problem.  Is there a reason we cannot update the driver again in monasca as well now?  IIRC the newer versions were unstable in some way (mem leaks or performance regressions).  If this has not been fixed then newer versions of kafka may still have good reason for being blocked.

Though I don't like it, we do have other pieces of software that are at least as good as depricated (psutil...)

Patch-set: 1