Update patch set 1
Patch Set 1:

Please look at the discussion on the ML:

http://lists.openstack.org/pipermail/openstack-dev/2016-December/108350.html
http://lists.openstack.org/pipermail/openstack-dev/2017-January/109741.html

and here: https://review.openstack.org/#/c/404878/

To recap:

We prove that Monasca issue is solved with recent version python-kafka. We provide patches for them to update their code to use the new version.

But their said, they have to double check them self...

And 1 month later, we have still no answer/input from monasca team and their -1 my patches without helpful comments.

This issue about this capping is 1 year old. 

The oslo.messaging driver for kafka is ready since 1 year and use the not deprecated lib. 

People on the ML said to use the not merged kafka driver in production.

I have raised the issue on the ML one more time at the end of November to be able to release oslo.messaging with kafka support for Ocata. We have already missed Mitaka and Newton...

I don't known how to move forward on this. I just see people frustrated to not be able to plug openstack on kafka.

The lib release is this week. And we are still blocked. I'm afraid this is for one more cycle...

sigh...

Patch-set: 1
Reviewer: Gerrit User 2813 <2813@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0