Update patch set 4
Patch Set 4:

hey Tony, any idea how to set kombu>=4.0.0/amqp>=2.0? i have a patch in oslo.messaging[1] that can support it but it can't be merged because it depends on requirements which also can't be merged because that fails if you bump requirements without merging my patch. i would like to not have to build a kombu3/kombu4 compatible patch

[1] https://review.openstack.org/#/c/409294/

Patch-set: 4
Label: Verified=0