Update patch set 6
Patch Set 6:

Andrey, understood.  Is your interface the SDK itself or OSC?  We have worked around underlying library issues in OSC to keep our users operating in the past, we planned to get this one in the next release (next week) if the SDK does not release a fix in the next few days.

Novaclient's issues at first glance do not appear to be related to quota set, but I have not dug in to those tests deeply yet to see why there are multiple networks to confuse the boot command.

Patch-set: 6