Update patch set 4
Patch Set 4:

What we can do is remove thr upper cap but don't update uc. Then you modify the patch to edit uc locally, merge the patch, and then we lift the g-r and require the new.release.

But not sure if that's a lot of fun. There is no way to keep the kombu3 code paths working somehow?

Patch-set: 4
Label: Verified=0