Add libdeprecation
We need this library in openstacksdk in order to deprecate various APIs
in an automated fashion. We will also use it in the future to keep up
with changes in REST APIs so we can properly move with them when needed.

PyPI: https://pypi.python.org/pypi/libdeprecation/
GitHub: https://github.com/briancurtin/deprecation

* Is the library actively maintained?

Yes. While it's new, it was created because it's necessary for
openstacksdk and one of the libraries it's replacing, pyrax. It has a
very small API footprint so active maintenance might not be that much,
but I'm the creator and I have a lot of need for this, as simple as it
may seem.

* Is the library good code?

Yes. It's straight-forward yet powerful, and requires no external
runtime dependencies.

* Is the library python 3 compatible?

Yes.

* Is the library license compatible?

Yes, Apache 2.0.

* Is the library already packaged in the distros we target (Ubuntu
latest / Fedora latest)?

No. It could be if this is truly needed, but I'm not entirely sure
why that is necessary for a client application. Please advise if this is
a hard requirement.

* Is the function of this library already covered by other libraries in
  global-requirements.txt?

No.

* Is the library required for OpenStack project or related dev or
  infrastructure setup?

Yes. OpenStack SDK has—and will throughout its lifetime—a need to
deprecate various parts of code as things change. As the versions of
REST APIs go through their own phases of becoming unsupported, so too
will entire packages built around those APIs. All the while, APIs
within the library will need to change for one reason or another,
and we will need to appropriately communicate that and curtail our
support over a period of time. This library does those things in an
automated fashion, improving communication while lessening the burden
to contributors.

A specific example of where this library is needed is
https://bugs.launchpad.net/python-openstacksdk/+bug/1657498

Change-Id: I8eacbfebcfce396b098fabc50861ecf50cf07acd