Update patch set 1
Patch Set 1:

Mehdi, I responded with my kafka-python findings from a couple rounds of testing today.  The library itself, just speaking to the deprecated objects, has some issues with memory and stability that I talk about on the mailing list.  If you need more detail there feel free to ask.

The new non-deprecated objects have not been tested because they require some level of engineering effort to maintain the reliability and performance that Monasca requires.  This is not simply a drop in upgrade.  Right now, if we were to use the new library in Monasca, we only have two options.  

1. We could work at the current performance level but suffer from data loss if we have trouble communicating to Kafka or if an API node were to crash.
2. Maintain the current data guarantees but run with far less performance (about 40x less throughput based on testing I did with the 0.9.5 client).

I don't think that anyone currently using Monasca wants to accept either of those options so we need to find a way to maintain the current data guarantees while using the async behaviour of the new client library.  That takes time and engineering effort to make that happen.  Is there anyone in the community willing to put in the effort to help build and test these new features at scale?

On another topic I'm curious what new features are you looking to get out of the new library.  Is there anything we can do to help you get the capabilities you want with the existing client?

Patch-set: 1