Update patch set 1
Patch Set 1: Code-Review+1

copy/pasting my note to ML.


is there a reason why you are against bumping Kafka up for OpenStack? it seems Monasca requires 0.9.5 and is content with it. but the oslo team has developed something that works well for the use case for the broader OpenStack ecosystem.

this seems like a better solution would be to just allow monasca to stay as is and the openstack requirements to progress rather than get blocked by a service that may or may not be deployed.

the same thing was done when we bumped elasticsearch requirements. there are multiple projects using elasticsearch. we didn't have anyone working on it in Ceilometer so rather than block the entire community. we let it proceed and we could catch up later if it was urgent.

Patch-set: 1