Update patch set 1
Patch Set 1:

as i understand it. the performance regression is resolved[1]. the blockers that monasca has currently is that they continue to have a requirement for synchronous functionality, which is not the use case oslo.messaging has. additionally, the memory requirements is not a leak but just the way synchronous api works. it takes the message and copies it so it never actually frees memory until roundtrip is complete... in asynchronous mode, it frees memory asap.

so yes, the flaws are there, but not for the oslo.messaging use case since they've move to support new async functionality.

i've summarised the knowledge dump sileht gave me so i may have missed something.

[1] http://lists.openstack.org/pipermail/openstack-dev/2016-December/108350.html

Patch-set: 1